% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Bibliodata.R
\docType{class}
\name{Bibliodata-class}
\alias{Bibliodata-class}
\title{The Bibliodata S4 R class}
\description{
An object of class Bibliodata allows to store several objects of class Datasettorep. Moreover, it allows to launch a shiny app which is designed to represent violinplots and heatmaps.
}
\section{Slots}{

\describe{
\item{\code{biblio}}{a list (class list) of objects of class Datasettorep.}
}}

