% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_ID.R
\name{conv_ID}
\alias{conv_ID}
\title{To obtain correspondances between 2 features nomenclatures}
\usage{
conv_ID(vec_id, convTab)
}
\arguments{
\item{vec_id}{a vector of class character chain. Each element is the name of a feature in a specific nomenclature whose the user wants to know his name in another nomenclature.}

\item{convTab}{a dataframe of two columns with in the first column the nomenclature of a large set of features (typically the set of all genes in a specific nomenclature such as Ensembl) 
In this column, features are in the nomenclature corresponding to those used as input of vec_id. The second column corresponds to the other nomenclature which will interest the user.}
}
\value{
a dataframe of two columns with in the first column the feature as provided by the user as input of the argument vec_id and in the second column the name of the feature in the other nomenclature.
}
\description{
Function that allows to give the name of features (here generally genes) to another nomenclature.
}
\examples{

conv_dataframe <- GTF_ensembl_to_conv_df(GTF_homo_sapiens_example())

conversion <- conv_ID(vec_id = c("KIT","GATA1","GATA2","NFE2"),convTab = conv_dataframe)

head(conversion)  

}
\author{
Elie Robert
}
