% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EG_RNAseq_metadata_example.R
\name{EG_RNAseq_metadata_example}
\alias{EG_RNAseq_metadata_example}
\title{EG_RNAseq_metadata_example}
\usage{
EG_RNAseq_metadata_example()
}
\description{
This function gives the path to an rds file containing a metadata dataframe corresponding to a bulk RNAseq experiment.
Samples are HEL (Human Erythroleukemia lineage) cells either control (3 samples), either carrying the ETO2::GLIS2 fusion, either carrying the ETO2::GLIS2 
without the zinc finger of the GLIS2 part of the ETO2-GLIS2 fusion or carrying the fusion ETO2::GLIS2 without the dNHR2 part of the ETO2-GLIS2 fusion  
(dimerization domain of the ETO2 part of the ETO2-GLIS2 fusion).
}
\examples{

EG_RNAseq_metadata_example()

}
