% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOXA2_RNAseq_count_matrix_example.R
\name{FOXA2_RNAseq_count_matrix_example}
\alias{FOXA2_RNAseq_count_matrix_example}
\title{FOXA2_RNAseq_count_matrix_example}
\usage{
FOXA2_RNAseq_count_matrix_example()
}
\description{
This function gives the path to an rds file containing a matrix corresponding to a bulk RNAseq experiment (GSE119931).
The data consists of 3 Wild Type (WT) samples and 3 Knock Outs (KO) of the transcription factor FOXA2.
}
\examples{

FOXA2_RNAseq_count_matrix_example()

}
