% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{get_counts}
\alias{get_counts}
\alias{get_counts,Datasettorep-method}
\alias{get_counts,Datasettorep}
\title{Accessor for the 'counts' slot of a Datasettorep object.}
\usage{
get_counts(object)

\S4method{get_counts}{Datasettorep}(object)
}
\arguments{
\item{object}{\code{Datasettorep} object}
}
\value{
matrix (class matrix) containing the counts data (rows corresponds to variables and columns to samples)
}
\description{
Accessors for the 'counts' slot of a Datasettorep object. The counts slot holds the count data as a matrix. The matrix (class matrix) contains values for each variables (rows) for each #' sample #' (column) get_counts allows to retrieve counts data from an object of class Datasettorep.
}
\examples{

require(readr)

expr_mat <- read_rds(FOXA2_RNAseq_count_matrix_example()) 

diff_results <- read_rds(FOXA2_differential_dataframe_example())

rownames(diff_results) <- diff_results$ID

## génération d un objet de class Datasettorep associé au jeu de données GLIS RNAseq 
FOXA2_RNAseq <- new("Datasettorep",
counts = expr_mat,
metadata = data.frame(condition = c(rep("WT",3),rep("FOXA2ko",3)),row.names = colnames(expr_mat)),
stattest = list("WT_vs_FOXA2ko" = diff_results),
colcond = "condition",
colsignificativity = "adj.P.Val",
significativity_symbols = list(c("***","**","NS"),c(0,0.01,0.05,1)),
description = "The normalized counts matrix (limma normalization) for this dataset 
(GSE119931; Balestrieri C et al, 2019, EMBO J) 
has been retrieved from the github of the package DecoupleR.
https://github.com/saezlab/decoupleR/tree/master/inst/extdata ")

get_counts(FOXA2_RNAseq)

}
\author{
Elie Robert
}
