% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{getcolcond}
\alias{getcolcond}
\alias{getcolcond,Datasettorep-method}
\alias{getcolcond,Datasettorep}
\title{Accessor for the 'colcond' slot of a \code{Datasettorep} object.}
\usage{
getcolcond(object)

\S4method{getcolcond}{Datasettorep}(object)
}
\arguments{
\item{object}{a \code{Datasettorep} object}
}
\value{
a character chain that gives the name of the column of the slot \code{metadata} that is used to specify the condition associated to each sample.
}
\description{
The slot 'colcond' is a character chain (R class character) that gives the name of the column of the slot \code{metadata} 
that should be use to specify the condition associated to each sample.
getcolcond allows to retrieve the name of the column of the metadata slot of a Datasettorep object that defines conditions associated to each condition.
}
\examples{
require(readr)

expr_mat <- read_rds(FOXA2_RNAseq_count_matrix_example()) 

diff_results <- read_rds(FOXA2_differential_dataframe_example())

rownames(diff_results) <- diff_results$ID

## génération d un objet de class Datasettorep associé au jeu de données GLIS RNAseq 
FOXA2_RNAseq <- new("Datasettorep",
counts = expr_mat,
metadata = data.frame(condition = c(rep("WT",3),rep("FOXA2ko",3)),row.names = colnames(expr_mat)),
stattest = list("WT_vs_FOXA2ko" = diff_results),
colcond = "condition",
colsignificativity = "adj.P.Val",
significativity_symbols = list(c("***","**","NS"),c(0,0.01,0.05,1)),
description = "The normalized counts matrix (limma normalization) for this dataset 
(GSE119931; Balestrieri C et al, 2019, EMBO J) 
has been retrieved from the github of the package DecoupleR.
https://github.com/saezlab/decoupleR/tree/master/inst/extdata ")


getcolcond(FOXA2_RNAseq)   
 
}
\author{
Elie Robert
}
