% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{getconv_df}
\alias{getconv_df}
\alias{getconv_df,Datasettorep-method}
\alias{getconv_df,Datasettorep}
\title{Accessor for the 'convdf' slot of a \code{Datasettorep} object.}
\usage{
getconv_df(object)

\S4method{getconv_df}{Datasettorep}(object)
}
\arguments{
\item{object}{a \code{Datasettorep} object}
}
\value{
a data.frame (the slot convdf of the \code{Datasettorep} object)that gives the corresponding name of each feature in another nomenclature when needed (for example ensembl gene ID and usual symbol gene name).
}
\description{
The slot convdf is a data.frame (R class data.frame) that gives the corresponding name of each feature in another nomenclature 
when needed (for example ensembl gene ID and usual symbol gene name). 
The first column should correspond to names in the the nomenclature used in the matrix given as value of the slot counts. 
Each character chain in this column should be unique. The second column contains corresponding names in the other  nomenclature. 
The same name can be found several times in this column.
}
\examples{

require(readr)
normalized_counts_EG <- read_rds(EG_RNAseq_count_matrix_example())

metadata_EG <- read_rds(EG_RNAseq_metadata_example())

DESeq2_EG <- read_rds(EG_RNAseq_differential_analysis_example())

conv_dataframe <- GTF_ensembl_to_conv_df(GTF_homo_sapiens_example())

EG_HEL_dataset  <- new("Datasettorep",
counts = normalized_counts_EG,
metadata = metadata_EG,
stattest = DESeq2_EG,
colcond = "condition",
colsignificativity = "padj",
significativity_symbols = list(c("***","**","NS"),c(0,0.01,0.05,1)),
conv_df = conv_dataframe,
description = "RNAseq of HEL cells (cell line derived from 
a patient suffering of an erythroleukemia). 
There is four conditions is this dataset : control HEL cells, HEL cells carrying the ETO2-GLIS2 
fusion (introduced by CRISPR-Cas9 editing), 
HEL cells carrying the ETO2-GLIS2 fusion but with a 
truncated C265G domain (zinc finger of the GLIS2 part of the protein), 
HEL cells carrying the ETO2 GLIS2 fusion 
but with a truncation of the dimerization domain of the ETO2 part of the protein
(Aid et al, Leukemia, 2022 ; Thirant et al, cancer Cell 2017). 
Quantification of reads has been performed #' with Salmon and subsequent analysis 
(Variance Stabilizing Transformation to obtain the normalized matrix 
and the differential expression analysis) have been performed by DESeq2)")

getconv_df(EG_HEL_dataset)

}
\author{
Elie Robert
}
