% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{getsignificativity_symbols}
\alias{getsignificativity_symbols}
\alias{getsignificativity_symbols,Datasettorep-method}
\alias{getsignificativity_symbols,Datasettorep}
\title{Accessor (getter) for the 'significativity_symbols' slot of a \code{Datasettorep} object.}
\usage{
getsignificativity_symbols(object)

\S4method{getsignificativity_symbols}{Datasettorep}(object)
}
\arguments{
\item{object}{a \code{Datasettorep} object}
}
\value{
a R list that gives the symbol associated to each level of significativity when printing violinplot.
}
\description{
getsignificativity_symbols allows to retrieve the slot significativity_symbols of a Datasettorep object.
This method allows to get the value of the 'significativity_symbols' slot of a \code{Datasettorep} object. 

significativity_symbols is a list (R class list) that gives the symbol associated to 
each level of significativity when printing violinplot. This list is composed of two elements : the first one is a character vector that gives the symbols that will be used 
The second vector is a numerical vector that specifies the limit values to choose what symbols will be use. 
For example, if a user provides the list list(c("*","NS"),c(0,0.05,1)) as a value of the parameter significativity_symbols, this means that differences associated
to a degree of significance between 0 and 0.05 will be represented with the symbol (character chain) * and differences associated to a degree of significance between 0.05 and 1 
should be ranked in ascending order.
}
\examples{

require(readr)

expr_mat <- read_rds(FOXA2_RNAseq_count_matrix_example()) 

diff_results <- read_rds(FOXA2_differential_dataframe_example())

rownames(diff_results) <- diff_results$ID

## génération d un objet de class Datasettorep associé au jeu de données GLIS RNAseq 
FOXA2_RNAseq <- new("Datasettorep",
counts = expr_mat,
metadata = data.frame(condition = c(rep("WT",3),rep("FOXA2ko",3)),row.names = colnames(expr_mat)),
stattest = list("WT_vs_FOXA2ko" = diff_results),
colcond = "condition",
colsignificativity = "adj.P.Val",
significativity_symbols = list(c("***","**","NS"),c(0,0.01,0.05,1)),
description = "The normalized counts matrix (limma normalization) for this dataset 
(GSE119931; Balestrieri C et al, 2019, EMBO J) 
has been retrieved from the github of the package DecoupleR.
https://github.com/saezlab/decoupleR/tree/master/inst/extdata ")

getsignificativity_symbols(FOXA2_RNAseq) 


}
\author{
Elie Robert
}
