% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{getstattest}
\alias{getstattest}
\alias{getstattest,Datasettorep-method}
\alias{getstattest,Datasettorep}
\title{Accessor for the 'stattest' slot of a \code{Datasettorep} object}
\usage{
getstattest(object)

\S4method{getstattest}{Datasettorep}(object)
}
\arguments{
\item{object}{\code{Datasettorep} object}
}
\value{
a R list. Each element of this list is a data.frame (class data.frame) 
with at least one column (whose name is given the slot colsignificativity)
that gives the level of significance (for instance, p-value or fdr) of the difference 
between two conditions for each feature.
Each element of the list corresponds to results of statistical analysis 
comparing features between two conditions
(conditions specified thanks to the data.frame corresponding to the slot \code{metadata})
}
\description{
The stattest slot is a list (class list). Each element of this list is a data.frame (class data.frame) 
with at least one column (whose name is given the slot colsignificativity) 
that gives the level of significance (for instance, p-value or fdr) 
of the difference between two conditions for each feature. 
That is why each element of the list should correspond to results of statistical analysis 
comparing features between two conditions 
(conditions specified thanks to the data.frame corresponding to the slot \code{metadata})
getstattest allows to retrieve the list of statistical results from an object of class Datasettorep.
}
\examples{

require(readr)

expr_mat <- read_rds(FOXA2_RNAseq_count_matrix_example()) 

diff_results <- read_rds(FOXA2_differential_dataframe_example())

rownames(diff_results) <- diff_results$ID

## génération d un objet de class Datasettorep associé au jeu de données GLIS RNAseq 
FOXA2_RNAseq <- new("Datasettorep",
counts = expr_mat,
metadata = data.frame(condition = c(rep("WT",3),rep("FOXA2ko",3)),row.names = colnames(expr_mat)),
stattest = list("WT_vs_FOXA2ko" = diff_results),
colcond = "condition",
colsignificativity = "adj.P.Val",
significativity_symbols = list(c("***","**","NS"),c(0,0.01,0.05,1)),
description = "The normalized counts matrix (limma normalization) for this dataset 
(GSE119931; Balestrieri C et al, 2019, EMBO J) 
has been retrieved from the github of the package DecoupleR.
https://github.com/saezlab/decoupleR/tree/master/inst/extdata ")

getstattest(FOXA2_RNAseq)

}
\author{
Elie Robert
}
