% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTF_homo_sapiens_example.R
\name{GTF_homo_sapiens_example}
\alias{GTF_homo_sapiens_example}
\title{GTF_homo_sapiens_example}
\usage{
GTF_homo_sapiens_example()
}
\description{
This function gives the path to a GTF file containing a metadata dataframe corresponding to a bulk RNAseq experiment. 
This GTF is a subset of the original GTF (Homo_sapiens.GRCh38.97) in order to reduce the size of files given to run the examples.
Samples are HEL (Human Erythroleukemia lineage) cells either control (3 samples), either carrying the ETO2::GLIS2 fusion, either carrying the ETO2::GLIS2 
without the zinc finger of the GLIS2 part of the ETO2-GLIS2 fusion or carrying the fusion ETO2::GLIS2 without the dNHR2 part of the ETO2-GLIS2 fusion  
(dimerization domain of the ETO2 part of the ETO2-GLIS2 fusion).
}
\examples{

GTF_homo_sapiens_example()

}
