% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheatmap_dataset.R
\name{pheatmap_dataset}
\alias{pheatmap_dataset}
\title{Function to generate pheatmaps}
\usage{
pheatmap_dataset(
  data,
  featurestorepresent,
  OrderSample,
  ColGroupMetadata,
  pathoutput,
  Cluster_Method = "complete",
  FontSizeCompHeatmap = 5,
  Title = "",
  toLog = TRUE,
  pdf_width = 15,
  pdf_height = 15,
  scalerow = FALSE,
  show_sample_name = FALSE,
  clustering_distance_samples = "euclidean",
  clustering_distance_features = "euclidean"
)
}
\arguments{
\item{data}{an object of class Datasettorep}

\item{featurestorepresent}{a character vector. Each element of the vector is a feature that the user wants to show in the heatmap.}

\item{OrderSample}{a character vector that gives the order in which the samples should be printed in the heatmap from the left to the right.}

\item{ColGroupMetadata}{a named character vector that provides the color that should be associated to each condition (groups of interest in the dataframe of metadata).}

\item{pathoutput}{a character chain that gives the path where the pdf containing the heatmap will be generated.}

\item{Cluster_Method}{character chain that specify the clustering method that should be apply by the algorithm. Only the following character chains work : c("none","complete","ward.D","ward.D2","single","average","mcquitty","median","centroid"). "none" means that no clustering will be  applied. Samples will be kept in the order specified.}

\item{FontSizeCompHeatmap}{a numeric value that specify the size of letters in the heatmap.}

\item{Title}{Character chain that gives the title of the heatmap.}

\item{toLog}{logical (TRUE or FALSE) if TRUE, then the values of the count matrix will be log2(x+1) transformed before clustering and generation of the heatmap. Otherwise, original values of the count matrix will be used for the clustering and the generation of the heatmap.}

\item{pdf_width}{Numeric value that specify the width of the output pdf that contains the heatmap.}

\item{pdf_height}{Numeric value that specify the height of the pdf that contains the heatmap.}

\item{scalerow}{logical value (TRUE or FALSE). If TRUE, then a z-transformation is applied on the features values before clustering and generation of the pheatmap. If FALSE, z-transformation is not applied.}

\item{show_sample_name}{logical value (TRUE or FALSE) If TRUE the name of each sample is printed in the output heatmap. If FALSE, names of samples are not printed in the output heatmap.}

\item{clustering_distance_samples}{a character chain that specify which types of distances between samples will be computed and used before clustering. The following distances can be choose :
"euclidean", "maximum", "manhattan", "canberra", "binary" ,"minkowski"}

\item{clustering_distance_features}{a character chain that specify which types of distances between features will be computed ands used before clustering. The following distances can be choose :
"euclidean", "maximum", "manhattan", "canberra", "binary" ,"minkowski"}
}
\description{
Function to generate heatmaps (based on the package pheatmap representations) based on a \code{Datasettorep} object.
}
\examples{

require(readr)

expr_mat <- read_rds(FOXA2_RNAseq_count_matrix_example()) 

diff_results <- read_rds(FOXA2_differential_dataframe_example())

rownames(diff_results) <- diff_results$ID

## génération d un objet de class Datasettorep associé au jeu de données GLIS RNAseq 
FOXA2_RNAseq <- new("Datasettorep",
counts = expr_mat,
metadata = data.frame(condition = c(rep("WT",3),rep("FOXA2ko",3)),row.names = colnames(expr_mat)),
stattest = list("WT_vs_FOXA2ko" = diff_results),
colcond = "condition",
colsignificativity = "adj.P.Val",
significativity_symbols = list(c("***","**","NS"),c(0,0.01,0.05,1)),
description = "The normalized counts matrix (limma normalization) for this dataset 
(GSE119931; Balestrieri C et al, 2019, EMBO J) 
has been retrieved from the github of the package DecoupleR.")

pheatmap_dataset(data = FOXA2_RNAseq, 
featurestorepresent = c("FOXA2","PLEKHN1","RHBDL2","FHL1","CLU","RBP4"),
OrderSample = c("WT","FOXA2ko"), 
ColGroupMetadata = c("WT" = "blue", "FOXA2ko" = "red"),
pathoutput = file.path(tempdir(), "test.pdf"),
Cluster_Method = "ward.D",
toLog = FALSE)

}
\author{
Elie Robert
}
