% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{setcolsignificativity<-}
\alias{setcolsignificativity<-}
\alias{setcolsignificativity<-,Datasettorep-method}
\alias{setcolsignificativity,Datasettorep,character}
\title{Accessor (setter) for the 'colsignificativity' slot of a \code{Datasettorep} object.}
\usage{
setcolsignificativity(object) <- value

\S4method{setcolsignificativity}{Datasettorep}(object) <- value
}
\arguments{
\item{object}{a \code{Datasettorep} object}

\item{value}{a character chain which corresponds to the column of each dataframe of the slot 'stattest' that gives the level 
of significance of the difference between two conditions for each feature.}
}
\value{
an object of class \code{Datasettorep} with eventually a modified value of the slot colsignificativity.
}
\description{
setcolsignificativity allows to give the name of the column that will be use to give the level of significativity of a statistical test between two conditions. 
This method allows to change the value of the 'colsignificativity' slot of a \code{Datasettorep} object.
The slot colsignificativity is a character chain (class character) that gives the name of the column in each data.frame 
of the slot \code{stattest} that provides the level of significance of the difference of 
a feature between two conditions.
}
\examples{

require(readr)

expr_mat <- read_rds(FOXA2_RNAseq_count_matrix_example()) 

diff_results <- read_rds(FOXA2_differential_dataframe_example())

rownames(diff_results) <- diff_results$ID

## génération d un objet de class Datasettorep associé au jeu de données GLIS RNAseq 
FOXA2_RNAseq <- new("Datasettorep",
counts = expr_mat,
metadata = data.frame(condition = c(rep("WT",3),rep("FOXA2ko",3)),row.names = colnames(expr_mat)),
stattest = list("WT_vs_FOXA2ko" = diff_results),
colcond = "condition",
colsignificativity = "P.Value",
significativity_symbols = list(c("***","**","NS"),c(0,0.01,0.05,1)),
description = "The normalized counts matrix (limma normalization) for this dataset 
(GSE119931; Balestrieri C et al, 2019, EMBO J) 
has been retrieved from the github of the package DecoupleR.
https://github.com/saezlab/decoupleR/tree/master/inst/extdata ")

getsignificativity_symbols(FOXA2_RNAseq)

setcolsignificativity(FOXA2_RNAseq) <- "adj.P.Val"

getsignificativity_symbols(FOXA2_RNAseq)

}
\author{
Elie Robert
}
