% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{setsignificativity_symbols<-}
\alias{setsignificativity_symbols<-}
\alias{setsignificativity_symbols<-,Datasettorep-method}
\alias{setsignificativity_symbols,Datasettorep,list}
\title{setter for the significativity_symbols slot}
\usage{
setsignificativity_symbols(object) <- value

\S4method{setsignificativity_symbols}{Datasettorep}(object) <- value
}
\arguments{
\item{object}{an object of class \code{Datasettorep} object.}

\item{value}{a R list. This list is composed of two elements : the first one is a character vector that gives the symbols that will be used 
The second vector is a numerical vector that specifies the limit values to choose what symbols will be use.
For example, if a user provides the list list(c("*","NS"),c(0,0.05,1)) as a value of the parameter significativity_symbols, this means that differences associated
to a degree of significance between 0 and 0.05 will be represented with the symbol (character chain) * and differences associated to a degree of significance between 0.05 and 1 
should be ranked in ascending order.}
}
\value{
an object of class Datasettorep with eventually a modified slot significativity_symbols if necessary.
}
\description{
setsignificativity_symbols allows to specify the list that give symbols associated to each level of significance (used in violinplot representations)
This method allows to change the value of the 'significativity_symbols' slot of a \code{Datasettorep} object.
significativity_symbols is a list (R class list) that gives the symbol associated to 
each level of significativity when printing violinplot. This list is composed of two elements : the first one is a character vector that gives the symbols that will be used 
The second vector is a numerical vector that specifies the limit values to choose what symbols will be use. 
For example, if a user provides the list list(c("*","NS"),c(0,0.05,1)) as a value of the parameter significativity_symbols, this means that differences associated
to a degree of significance between 0 and 0.05 will be represented with the symbol (character chain) * and differences associated to a degree of significance between 0.05 and 1 
should be ranked in ascending order.
}
\examples{

require(readr)

expr_mat <- read_rds(FOXA2_RNAseq_count_matrix_example()) 

diff_results <- read_rds(FOXA2_differential_dataframe_example())

rownames(diff_results) <- diff_results$ID

## génération d un objet de class Datasettorep associé au jeu de données GLIS RNAseq 
FOXA2_RNAseq <- new("Datasettorep",
counts = expr_mat,
metadata = data.frame(condition = c(rep("WT",3),rep("FOXA2ko",3)),row.names = colnames(expr_mat)),
stattest = list("WT_vs_FOXA2ko" = diff_results),
colcond = "condition",
colsignificativity = "P.Value",
significativity_symbols = list(c("*","NS"),c(0,0.05,1)),
description = "The normalized counts matrix (limma normalization) for this dataset 
(GSE119931; Balestrieri C et al, 2019, EMBO J) 
has been retrieved from the github of the package DecoupleR.
https://github.com/saezlab/decoupleR/tree/master/inst/extdata ")

getsignificativity_symbols(FOXA2_RNAseq)

setsignificativity_symbols(FOXA2_RNAseq) <- list(c("***","**","NS"),c(0,0.01,0.05,1))

getsignificativity_symbols(FOXA2_RNAseq)

}
\author{
Elie Robert
}
