% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{subset_bibliodata}
\alias{subset_bibliodata}
\alias{subset_bibliodata,Bibliodata,character-method}
\alias{subset_bibliodata,Bibliodata,character}
\alias{subset_bibliodata,Bibliodata,numeric-method}
\alias{subset_bibliodata,Bibliodata,numeric}
\title{getter function that allows to subset a Bibliodata object}
\usage{
subset_bibliodata(object, sel)

\S4method{subset_bibliodata}{Bibliodata,character}(object, sel)

\S4method{subset_bibliodata}{Bibliodata,numeric}(object, sel)
}
\arguments{
\item{object}{an object of class Bibliodata the user wants to subset.}

\item{sel}{Character vector either of class character or of class numeric. It corresponds either to a set of names of elements of class Datasettorep contained in the Bibliodata object or to a set #' #' of index (positions) of Datasettorep objects in the Bibliodata object.}
}
\value{
an object of class Bibliodata.
}
\description{
subset_bibliodata allows to subset a Bibliodata object to obtain a smaller Bibliodata object.
This function is a getter function that allows to subset a Bibliodata object and generate a new object of the same class. This object is a smaller bibliodata object that corresponds to a subset of #' the original bibliodata object.
}
\examples{

require(readr)

expr_mat <- read_rds(FOXA2_RNAseq_count_matrix_example()) 

diff_results <- read_rds(FOXA2_differential_dataframe_example())

rownames(diff_results) <- diff_results$ID

FOXA2_RNAseq <- datasettorep(counts = expr_mat, 
metadata = data.frame(condition = c(rep("WT",3),rep("FOXA2ko",3)),row.names = colnames(expr_mat)), 
colcond = "condition", 
colsignificativity = "adj.P.Val" , 
stattest = list("WT_vs_FOXA2ko" = diff_results),
significativity_symbols = list(c("***","**","NS"),c(0,0.01,0.05,1)),
description = "The normalized counts matrix (limma normalization) for this dataset 
(GSE119931; Balestrieri C et al, 2019, EMBO J) 
has been retrieved from the github of the package DecoupleR.
https://github.com/saezlab/decoupleR/tree/master/inst/extdata")


normalized_counts_EG <- read_rds(EG_RNAseq_count_matrix_example())

metadata_EG <- read_rds(EG_RNAseq_metadata_example())

DESeq2_EG <- read_rds(EG_RNAseq_differential_analysis_example())

conv_dataframe <- GTF_ensembl_to_conv_df(GTF_homo_sapiens_example())

EG_HEL_dataset  <- new("Datasettorep",
counts = normalized_counts_EG,
metadata = metadata_EG,
stattest = DESeq2_EG,
colcond = "condition",
colsignificativity = "padj",
significativity_symbols = list(c("***","**","NS"),c(0,0.01,0.05,1)),
conv_df = conv_dataframe,
description = "RNAseq of HEL cells (cell line derived from 
a patient suffering of an erythroleukemia). 
There is four conditions is this dataset : control HEL cells, HEL cells carrying the ETO2-GLIS2 
fusion (introduced by CRISPR-Cas9 editing), 
HEL cells carrying the ETO2-GLIS2 fusion but with a 
truncated C265G domain (zinc finger of the GLIS2 part of the protein), 
HEL cells carrying the ETO2 GLIS2 fusion 
but with a truncation of the dimerization domain of the ETO2 part of the protein
(Aid et al, Leukemia, 2022 ; Thirant et al, cancer Cell 2017). 
Quantification of reads has been performed #' with Salmon and subsequent analysis 
(Variance Stabilizing Transformation to obtain the normalized matrix 
and the differential expression analysis) have been performed by DESeq2)")

example_bibliodata <- bibliodata(biblio = list("FOXA2_RNAseq" = FOXA2_RNAseq,
"EG_HEL" = EG_HEL_dataset))


str(subset_bibliodata(example_bibliodata,"FOXA2_RNAseq"))

str(subset_bibliodata(example_bibliodata,1))

}
\author{
Elie Robert
}
