% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlnplot_dataset.R
\name{vlnplot_dataset}
\alias{vlnplot_dataset}
\title{Method to generate violinplots based on a \code{Datasettorep} object.}
\usage{
vlnplot_dataset(
  data,
  genetorepresent,
  ordercondition,
  color_to_fill,
  ToLog,
  yposminbar = 2,
  yposmaxbar = 8,
  motif = NULL,
  size_text_axis = 19,
  size_title_axis = 19,
  x_axis_title = 19,
  show_significativity_stars = FALSE,
  dot_size = 0.7,
  dot_color = NA,
  bin_width = 0.037,
  size_stat_summary = 2,
  color_stat_summary = NA,
  box_width = 0.4,
  thickness_median_line = 3
)
}
\arguments{
\item{data}{a \code{Datasettorep} object}

\item{genetorepresent}{vector of class character. Each element in the vector corresponds to the name of a feature the user wants to represent (violinplot representation)}

\item{ordercondition}{vector of class character. Each element corresponds to the name of a condition in the dataset that the user wants to represent. 
Violins associated to each condition will be printed in the same order as in the vector.}

\item{color_to_fill}{vector of class character. Each color corresponds to the color of the violin at the same position in the violinplot representation as the position in the vector.}

\item{ToLog}{(logical) if the counts matrix should be log2(x+1) transformed.}

\item{yposminbar}{(numeric) a numeric value that specify where should be the horizontal bar (associated to symbols of significance) the lower in the graph compared to the higher point 
(maximum value) for a given feature.}

\item{yposmaxbar}{(numeric) a numeric value that specify where (how higher) should be the horizontal bar (associated to symbols of significance) the higher in the graph compared to the higher point
(maximum value) for a given feature.}

\item{motif}{(character chain) this character chain corresponds to the condition that is compared to others in the case where the parameter show_significativity_stars is set to TRUE.}

\item{size_text_axis}{(numeric) size of text associated to graduations}

\item{size_title_axis}{(numeric) size of the axis title.}

\item{x_axis_title}{(character chain). Name of the x axis.}

\item{show_significativity_stars}{(logical) whether or not display symbols of significativity}

\item{dot_size}{(numeric) The diameter of the dots relative to binwidth}

\item{dot_color}{(character chain) color of points}

\item{bin_width}{(numeric) In a dot plot, the width of a dot corresponds to the bin width (or maximum width, depending on the binning algorithm)
and dots are stacked, with each dot representing one observation.}

\item{size_stat_summary}{(numeric) size of the mean symbol for all the violin in the representation.}

\item{color_stat_summary}{(character) the color of the mean symbol.}

\item{box_width}{(numeric) The width of the box of the violins}

\item{thickness_median_line}{(numeric) The thickness of the median line.}
}
\description{
Method to generate violinplots based on a \code{Datasettorep} object.
}
\examples{

require(readr)

expr_mat <- read_rds(FOXA2_RNAseq_count_matrix_example()) 

diff_results <- read_rds(FOXA2_differential_dataframe_example())

str(diff_results)

rownames(diff_results) <- diff_results$ID

## génération d un objet de class Datasettorep associé au jeu de données GLIS RNAseq 
FOXA2_RNAseq <- new("Datasettorep",
counts = expr_mat,
metadata = data.frame(condition = c(rep("WT",3),rep("FOXA2ko",3)),row.names = colnames(expr_mat)),
stattest = list("WT_vs_FOXA2ko" = diff_results),
colcond = "condition",
colsignificativity = "adj.P.Val",
significativity_symbols = list(c("***","**","NS"),c(0,0.01,0.05,1)),
description = "The normalized counts matrix (limma normalization) for this dataset 
(GSE119931; Balestrieri C et al, 2019, EMBO J) 
has been retrieved from the github of the package DecoupleR.",
conv_df = data.frame())

vlnplot_dataset(data = FOXA2_RNAseq,
genetorepresent = c("FOXA2","PLEKHN1","RHBDL2","FHL1","CLU","RBP4"),
ordercondition = c("WT","FOXA2ko"),
color_to_fill = c("WT" = "blue","FOXA2ko" = "red"),
show_significativity_stars = TRUE,
ToLog = TRUE,
motif = "FOXA2ko"
)

}
\author{
Elie Robert
}
